+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## [0.1.1]
+- add reveal template folder command
+- open template folder `with new instance`
+- fix wrong capture with the pattern `typedef struct {}`
 ## [0.1.0]
-
 - Initial release